fix(runtime): show child workflow command shape
